Now we're getting there..
Updated to include two standard callbacks. OnMetadata and onCuePoint.
Some streams will fail to play if there isn't an onMetaData method
specified in a client object. This has been included by default with the
option to define a standard listener to get the onMetaData callback or
assign a client object. Very handy when wanting to quickly build a
player and not worry about boiler plate code. This also stores the
metadata for reference at any time.